requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.2
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.3.0.23
cached-property==1.5.2
celery==3.1.26.post2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
decorator==5.1.0
Django==3.2.9
django-celery==3.3.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.1
incremental==21.3.0
ipython==5.1.0
jsonfield==2.0.2
kombu==3.0.37
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==1.0.18
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyOpenSSL==21.0.0
pytz==2021.3
requests==2.26.0
service-identity==21.1.0
simplegeneric==0.8.1
six==1.16.0
sqlparse==0.4.2
traitlets==5.1.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
zope.interface==5.4.0
